started refactoring the createUserSection compontnt. Made a separate customhook and request function to it
