+++ frontend/src/components/CreateUserSection.tsx
@@ -1,37 +1,23 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useState } from "react";
-import { Users } from "../services/api/usersTypes";
+import useCreateNewUser from "../hooks/useCreateNewUser";
 
 export const CreateUserSection = () => {
 	const [newUserName, setNewUserName] = useState<string>("");
 	const [newUserEmail, setNewUserEmail] = useState<string>("");
-	const queryClient = useQueryClient();
 
-	const createMutation = useMutation({
-		mutationFn: async (newUser: Pick<Users, "name" | "email">) => {
-			const response = await fetch("http://localhost:4000/users", {
-				method: "POST",
-				headers: {
-					"Content-Type": "application/json",
-				},
-				body: JSON.stringify(newUser),
-			});
-			const data = response.json();
-			console.log(data);
-		},
-		onSuccess: () => {
-			queryClient.invalidateQueries({
-				queryKey: ["users"],
-			});
-		},
-	});
+	const createMutation = useCreateNewUser();
 
-	const handleCreateUsers = () => {};
+	const handleCreateUsers = () => {
+		createMutation.mutate({ name: newUserName, email: newUserEmail });
+		setNewUserName("");
+		setNewUserEmail("");
+	};
 
 	return (
 		<section>
 			<h3>Create a User Here!</h3>
 			<label htmlFor='name'>Name:</label>
+			<br />
 			<input
 				type='text'
 				id='name'
@@ -42,6 +28,7 @@ export const CreateUserSection = () => {
 			/>{" "}
 			<br />
 			<label htmlFor='email'>Email:</label>
+			<br />
 			<input
 				type='text'
 				id='email'
