+++ frontend/src/services/api/usersRequests.ts
@@ -4,3 +4,13 @@ export const fetchUsers = async () => {
 	const response = await fetch("http://localhost:4000/users");
 	return (await response.json()) as Users[];
 };
+
+export const createUser = async (newUser: Pick<Users, "name" | "email">) => {
+	return fetch("http://localhost:4000/users", {
+		method: "POST",
+		headers: {
+			"Content-Type": "application/json",
+		},
+		body: JSON.stringify(newUser),
+	});
+};
